Moving shared utilities to a new package.
