+++ bigtable-dataflow-parent/bigtable-hbase-dataflow/src/main/java/com/google/cloud/bigtable/batch/util/ByteStringUtil.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.bigtable.dataflow;
+package com.google.cloud.bigtable.batch.util;
 
 import com.google.bigtable.repackaged.com.google.cloud.util.ByteStringer;
 import com.google.bigtable.repackaged.com.google.protobuf.ByteString;
