+++ bigtable-dataflow-parent/bigtable-hbase-dataflow/src/main/java/com/google/cloud/bigtable/batch/util/package-info.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright (C) 2015 Google Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+/**
+ * <p>
+ * Google Cloud Bigtable Utilities used by Dataflow.
+ * </p>
+ * <p>
+ * Google Cloud Bigtable offers you a fast, fully managed, massively scalable NoSQL database service
+ * that's ideal for web, mobile, and Internet of Things applications requiring terabytes to
+ * petabytes of data. Unlike comparable market offerings, Cloud Bigtable doesn't require you to
+ * sacrifice speed, scale, or cost efficiency when your applications grow. Cloud Bigtable has been
+ * battle-tested at Google for more than 10 years--it's the database driving major applications such
+ * as Google Analytics and Gmail.
+ * </p>
+ */
+package com.google.cloud.bigtable.batch.util;
